pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "intellireading-api_server"
dynamic = ["version"]
description = 'Intellireading-api_server is a RESTful API server that provides a way to interact with the Intellireading-cli library.'
readme = "README.md"
requires-python = ">=3.13"
license = "GPL-3.0"
keywords = ["kepub","epub", "calibre", "ebook","kindle","kobo","speed-reading", "dyslexia", "adhd"]
authors = [
  { name = "Hugo Batista"},
]
classifiers = [
  "Operating System :: OS Independent",
  "Topic :: Software Development",
  "Topic :: Text Processing :: Markup :: XML",
  "Topic :: Education :: Computer Aided Instruction (CAI)",
  "Topic :: Utilities",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "Intended Audience :: Developers",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Education",
  "Intended Audience :: Science/Research",
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = [
  "Werkzeug>=3.0.4",
  "python_multipart>=0.0.9",
  "fastapi>=0.115.6",
  "uvicorn>=0.34.0",
  "opentelemetry-distro>=0.50b0",
  "opentelemetry-instrumentation-logging>=0.50b0",
  "opentelemetry-instrumentation-fastapi>=0.50b0",
  "opentelemetry-exporter-otlp>=1.29.0",
  "aiohttp>=3.10.5",
  "intellireading-cli>=2.1.0",
]
[tool.hatch.metadata]
allow-direct-references = true


[project.scripts]
api-server = "intellireading.api_server:entrypoint"

[project.urls]
Documentation = "https://go.hugobatista.com/gh/intellireading-api_server#readme"
Issues = "https://go.hugobatista.com/gh/intellireading-api_server/issues"
Source = "https://go.hugobatista.com/gh/intellireading-api_server"

[tool.hatch.version]
path = "src/intellireading/api_server/__about__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/intellireading"]
exclude = ["tests"]

[tool.hatch.build.targets.sdist]
packages = ["src/intellireading"]
exclude = ["tests"]

[tool.hatch.envs.default]
# using uv for faster installs
installer = "uv"

dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
  "httpx"
  #"intellireading-cli @ {root:parent:uri}/intellireading-cli",
]

# Pre-install dependencies that are not available on PyPI
pre-install-commands = [
   "python -c \"import glob, subprocess; [subprocess.run(['uv', 'pip', 'install', whl]) for whl in glob.glob('.dependencies/*.whl')]\"",
]

[tool.hatch.envs.default.env-vars]
API_SERVER_API_KEY = "devapikey"

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests} -v"
test-cov = "coverage run -m pytest {args:tests} -v"
cov-report = [
  "- coverage combine",
  "coverage report",
  "coverage html",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.13"]

[tool.hatch.envs.lint]
detached = true
# using uv for faster installs
installer = "uv"

dependencies = [
  "black>=23.1.0",
  "mypy>=1.0.0",
  "ruff>=0.0.243",
]
[tool.hatch.envs.lint.scripts]
typing = "mypy --ignore-missing-imports --install-types --non-interactive {args:src/intellireading tests}"
style = [
  "ruff check {args:.}",
  "black --check --diff {args:.}",
]
fmt = [
  "black {args:.}",
  "ruff --fix {args:.}",
  "style",
]
all = [
  "style",
  "typing",
]

[tool.black]
target-version = ["py37"]
line-length = 100
skip-string-normalization = true

[tool.ruff]
target-version = "py37"
line-length = 100
lint.select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
lint.ignore = [
  # Ignore Import block is un-sorted or un-formatted
  "I001",
  # Ignore function call in argument defaults
  "B008",
]
lint.unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.lint.isort]
known-first-party = ["intellireading"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.coverage.run]
source_pkgs = ["intellireading", "tests"]
branch = true
parallel = true
omit = [
  "src/intellireading/api_server/__about__.py",
]

[tool.coverage.paths]
intellireading = ["src/intellireading", "*/intellireading-api_server/src/intellireading"]
tests = ["tests", "*/intellireading-api_server/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
